+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ## 2.8 branch
 ### 2.8.1
 * improved the recovery of the information during the addition of a video;
-* updated for MeCms 2.16.1.
+* updated for MeCms 2.16.1;
+* added tests for `AppController` class.
 
 ### 2.8.0
 * the cells that act as widgets now have "Widgets" in the name, for the classes